# parallel_copy