--- a/src/utils.rs
+++ b/src/utils.rs
@@ -428,6 +428,25 @@ fn packed_eq_poly<F: Field, EF: ExtensionField<F>>(
     EF::ExtensionPacking::from_ext_slice(&buffer)
 }
 
+pub fn parallel_copy<A: Copy + Send + Sync>(src: &[A], dst: &mut [A]) {
+    #[cfg(feature = "parallel")]
+    if src.len() < 1 << 15 {
+        // sequential copy
+        dst.copy_from_slice(src);
+    } else {
+        assert_eq!(src.len(), dst.len());
+        let chunk_size = src.len() / rayon::current_num_threads().max(1);
+        dst.par_chunks_mut(chunk_size)
+            .zip(src.par_chunks(chunk_size))
+            .for_each(|(d, s)| {
+                d.copy_from_slice(s);
+            });
+    }
+
+    #[cfg(not(feature = "parallel"))]
+    dst.copy_from_slice(src);
+}
+
 #[cfg(test)]
 mod tests {
     use p3_baby_bear::BabyBear;
@@ -438,6 +457,20 @@ mod tests {
 
     type F = BabyBear;
     type EF4 = BinomialExtensionField<F, 4>;
+
+    #[test]
+    fn test_parallel_copy() {
+        let src = (0..(1 << 25) + 7).map(F::from_u64).collect::<Vec<_>>();
+        let mut dst_seq = F::zero_vec(src.len());
+        let time = std::time::Instant::now();
+        dst_seq.copy_from_slice(&src);
+        println!("Sequential copy took: {:?}", time.elapsed());
+        let mut dst_parallel = F::zero_vec(src.len());
+        let time = std::time::Instant::now();
+        parallel_copy(&src, &mut dst_parallel);
+        println!("Parallel copy took: {:?}", time.elapsed());
+        assert_eq!(dst_seq, dst_parallel);
+    }
 
     #[test]
     fn test_eval_eq_functionality() {