To get an idea an idea of the speedup:

```
RUSTFLAGS='-C target-cpu=native' cargo test --release --package whir-p3 --lib -- utils::tests::test_parallel_copy --exact --show-output
```

Running `main.rs`, and analyzing the time spent on the 5 invocations of `copy_across_coeffs`, on my machine:
| Before    | After |
| -------- | ------- |
| 42.7ms  | 9.84ms    |
| 7.59ms | 2.25ms     |
| 424µs    | 181µs    |
| 84.0µs    | 75.9µs    |
| 736µs    | 743µs    |